requirements.txt
aiochclient
aiohttp
APScheduler
dbutils
happybase~=1.2.0
loguru~=0.6.0
lxml
matplotlib
motor~=3.0.0
numpy
openpyxl
pandahouse
pandas~=1.4.1
pycryptodome
PyMySQL~=0.9.3
pyspark~=3.2.1
redis~=4.2.2
requests~=2.27.1
selenium
telethon
toolz
urllib3
wordcloud
aiomysql~=0.0.22
cx_Oracle
cx_Oracle_async
grpcio
